ADR-971 Fix test broken in merge
